Move generic reusable widgets into picard.ui.widgets module
